refactor service functions into shared module
